Removal of node-fibers wrapper
The wrapper script is no longer needed because we're not depending on
any LD_PRELOAD magic anymore. A few build peculiarities had to be taken
care of, and the bootstrapping in coroutine.cc was changed slightly.
Also SJLJ in libcoro seems to hang on Linux for me, so I'm specifying
UCONTEXT (old method) for Linux builds now.