Fix segfault with inherited fibers
If you inherit a fiber in a prototype there may be a segfault because
the `started` accessor gets an invalid `this`. I can't figure out how to
fix this properly with the Holder() function like I can for methods.. so
instead of segfaulting I'm just going to return `undefined` here. You
shouldn't be using fibers in a prototype chain anyway.

Fixes gh-44 (maybe?)