Make Future.wrap deal with a fn.length of 0
I'm not sure if this is the correct way to handle things by
default, or even if there is a correct way to handle
things by default.  My use case here is with crypto.randomBytes.
In node, crypto.randomBytes.length returns 0, which causes idx to
be -1.  This fixes the wrapping of crypto.randomBytes, with no
side-effects in my code, but I'm not sure it is correct. It assumes
that the callback function is the last argument, and the user is
providing all arguments other than the callback.  That seems like
a reasonable default to me.