Don't rename node module for different archs
Turns out you can't load a module on Windows that has been renamed from
what it was called in the dllexport. So if you have fibers.node, and
rename it to fibers2.node, `require('fibers2') will fail with "unknown
error". I'm not sure if this is an inherent problem with dynamic
libraries in Windows or just Node being pedantic.

As a workaround just make a new directory for each platform and arch
with only one file, "fibers.node".