Detect failure to create coro stacks
If you are creating a lot of fibers (2000~) on a 32-bit system it's very
possible you will run out of addressable memory. This detects a failed
fiber creation and will throw a JS exception. However, if you are that
close to the memory ceiling it's likely that v8 is about to crash as
well.