Support for io.js and improved stack guard
In some pathological cases you could corrupt the stack using fibers and
cause a segfault. There seem to be some changes to how the stack guard
works from v8 3.14.x (node 0.10.x) to 3.28.x (node 0.12.x). This was
causing a lot of confusion so I brushed the problem under the rug a
little bit.