Fix segfault on exit when unwound fibers exist
Newer versions of v8 raise an exception when you try to delete an
isolate while other threads still have an active entry. This causes node
to crash instead of exiting gracefully. It's not too big of a deal since
we were in the process of terminating and no more user code could run
anyway, but it is still confusing.

Also this updates the unit tests to detect unusual exit codes and report
them as error. And also to exit with an error code when a test fails.

Fixes #223