Use v8 symbols for thread locals on musl
musl allocates thread locals round-robin from a hard-capped pool of 128
keys. They seed the search from the current stack pointer, which is
basically random. Most of the time the v8 thread locals will be
contiguous, however a little less than 1% of the time they will not be.
Furthermore, half the time the fibers coro key will greater than the v8
thread keys, which messes with our existing search.

This patch uses internal v8 symbols to grab the thread locals. This
method doesn't work on distributions which strip internal symbols so we
gate it behind a #define.