Use `process.versions.modules` to select binary
Node 6.5.0 updated the v8 version in the middle of a release cycle which
has never been done before but technically doesn't break ABI
compatibility. But then fibers tries to select a new binary which wasn't
packaged in npm and everyone gets mad at me. So now we're using
`process.versions.modules` which should only be updated for major
releases.